Views y permisos
Views y permisos

+++ src/services/session.service.js
@@ -251,28 +251,71 @@ export default class SessionService {
         return response;
     };
 
-    async deleteUserService(uid, cid, role) {
+    async deleteUserService(uid, role) {
         let response = {};
         try {
-            const resultDAO = await this.sessionDAO.deleteUser(uid);
-            if (resultDAO.status === "error") {
+            let cid;
+            let email;
+            let name;
+            const userInfo = await this.getUserService(uid)
+            if (userInfo.statusCode === 500) {
                 response.statusCode = 500;
-                response.message = resultDAO.message;
-            } else if (resultDAO.status === "not found user") {
+                response.message = userInfo.message;
+            } else if (userInfo.statusCode === 404) {
                 response.statusCode = 404;
-                response.message = `No se encontró ninguna cuenta con este ID, ${uid}.`;
-            } else if (resultDAO.status === "success") {
-                const deleteCart = await this.cartService.deleteCartService(cid);
-                const deleteUserProducts = await this.productsService.deleteAllPremiumProductService(uid, uid, role);
-                if (deleteCart.statusCode === "error" || deleteUserProducts.statusCode === "error") {
+                response.message = `No se encontró ningún usuario con el Email, Nombre o ID, ${identifier}.`;
+            } else if (userInfo.statusCode === 200) {
+                cid = userInfo.result.cart;
+                email = userInfo.result.email;
+                name = userInfo.result.first_name;
+                const resultDAO = await this.sessionDAO.deleteUser(uid);
+                if (resultDAO.status === "error") {
                     response.statusCode = 500;
-                    response.message = "Error al eliminar la cuenta: " + deleteCart.message || deleteUserProducts.message;
-                } else if (deleteCart.statusCode === 200 && deleteUserProducts.statusCode === 404) {
-                    response.statusCode = 200;
-                    response.message = "Cuenta eliminada exitosamente. No se encontraron productos asociados a la cuenta.";
-                } else if (deleteCart.statusCode === 200 && deleteUserProducts.statusCode === 200) {
-                    response.statusCode = 200;
-                    response.message = "Cuenta eliminada exitosamente. " + deleteUserProducts.message;
+                    response.message = resultDAO.message;
+                } else if (resultDAO.status === "not found user") {
+                    response.statusCode = 404;
+                    response.message = `No se encontró ninguna cuenta con este ID, ${uid}.`;
+                } else if (resultDAO.status === "success") {
+                    const deleteCart = await this.cartService.deleteCartService(cid);
+                    const deleteUserProducts = await this.productsService.deleteAllPremiumProductService(uid, uid, role);
+                    let errorCorreo;
+                    if (role === "admin") {
+                        let html = `
+                        <table cellspacing="0" cellpadding="0" width="100%">
+                            <tr>
+                                <h1>MECATRON</h1>
+                            </tr>
+                            <tr>
+                                <td style="text-align: center;">
+                                    <h2 style="font-size: 24px; margin: 0;">Notificación de eliminación de cuenta</h2>
+                                    <p style="font-size: 16px;">
+                                        Estimado ${name}, lamentamos informarte que tu cuenta ha sido eliminada por el administrador. Esta acción fue necesaria para mantener la calidad y seguridad de nuestra plataforma, si consideras que cometimos un error, tienes alguna pregunta o necesitas asistencia, no dudes en ponerte en contacto con nuestro equipo de soporte.
+                                    </p>
+                                    <p style="font-size: 16px; font-weight: bold;">
+                                        - Fecha y hora de eliminación: ${new Date().toLocaleDateString()} - ${new Date().toLocaleTimeString()}
+                                    </p>
+                                    <p style="font-size: 16px;">
+                                        Gracias por tu comprensión.
+                                    </p>
+                                </td>
+                            </tr>
+                        </table>`
+                        const resultSendMail = await this.mail.sendMail(email, "Notificación de eliminación de cuenta", html);
+                        if (resultSendMail.rejected && resultSendMail.rejected.length > 0) {
+                            errorCorreo = true;
+                            errorCorreo.message = "La cuenta, se ha eliminado pero ocurrio un error al intantar enviar el correo de notificación al usuario."
+                        }
+                    }
+                    if (deleteCart.statusCode === "error" || deleteUserProducts.statusCode === "error" || errorCorreo === true) {
+                        response.statusCode = 500;
+                        response.message = "Error al eliminar la cuenta: " + deleteCart.message || deleteUserProducts.message || errorCorreo.message
+                    } else if (deleteCart.statusCode === 200 && deleteUserProducts.statusCode === 404) {
+                        response.statusCode = 200;
+                        response.message = "Cuenta eliminada exitosamente. No se encontraron productos asociados a la cuenta.";
+                    } else if (deleteCart.statusCode === 200 && deleteUserProducts.statusCode === 200) {
+                        response.statusCode = 200;
+                        response.message = "Cuenta eliminada exitosamente. " + deleteUserProducts.message;
+                    }
                 }
             };
         } catch (error) {
@@ -281,5 +324,4 @@ export default class SessionService {
         };
         return response;
     };
-
 };
